refactor: Calculator, Parser, Validator 클래스 분리
