+++ src/main/java/calculator/Validator.java
@@ -1,0 +1,27 @@
+package calculator;
+
+public class Validator {
+    public static int validate(final String str) {
+        if (str.isEmpty()) return 0;
+
+        if (isNotDigit(str))
+            throw new IllegalArgumentException("[ERROR] 지정되지 않은 구분자입니다.");
+
+        int num = Integer.parseInt(str);
+        if (isNegative(num))
+            throw new IllegalArgumentException("[ERROR] 음수는 입력할 수 없습니다.");
+
+        return num;
+    }
+
+    private static boolean isNotDigit(final String str) {
+        for (char c : str.toCharArray()) {
+            if (!Character.isDigit(c)) return true;
+        }
+        return false;
+    }
+
+    private static boolean isNegative(int num) {
+        return num < 0;
+    }
+}
